requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
idna==3.4
oauthlib==3.2.2
sqlparse==0.4.4
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.4
gunicorn==21.2.0
packaging==23.1
django-storages==1.13.2
azure-core==1.26.3
azure-storage-blob==12.14.1
